pyproject.toml
[project]
name = "gs-docs"
version = "0.1.0"
description = "Unified documentation for projects"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dataclasses-json>=0.6.7",
    "httpx>=0.27.2",
    "mkdocs>=1.6.1",
    "pendulum>=3.0.0",
    "pyyaml>=6.0.2",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.3",
]

[tool.ruff.format]
quote-style = "single"
